Tested android example app
