+++ IapExample/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.dooboolab.iapexample",
+    package = "com.iapexample",
 )
 
 android_resource(
     name = "res",
-    package = "com.dooboolab.iapexample",
+    package = "com.iapexample",
     res = "src/main/res",
 )
