--- IapExample/android/app/src/main/java/com/dooboolab/iapexample/MainActivity.java
@@ -1,15 +1,0 @@
-package com.dooboolab.iapexample;
-
-import com.facebook.react.ReactActivity;
-
-public class MainActivity extends ReactActivity {
-
-    /**
-     * Returns the name of the main component registered from JavaScript.
-     * This is used to schedule rendering of the component.
-     */
-    @Override
-    protected String getMainComponentName() {
-        return "IapExample";
-    }
-}
