+++ IapExample/android/gradle.properties
@@ -17,7 +17,5 @@
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
-MYAPP_RELEASE_STORE_FILE=dooboo_test.jks
-MYAPP_RELEASE_KEY_ALIAS=key0
-MYAPP_RELEASE_STORE_PASSWORD=dooboolab
-MYAPP_RELEASE_KEY_PASSWORD=dooboolab
+android.useAndroidX=true
+android.enableJetifier=true
